+++ docker-entry.bash
@@ -4,5 +4,5 @@
 ! test -e /trans/trans.keystore && /trans/make-keystore.bash
 
 # Runt the server
-java -jar /trans/target/scala-2.11/trans-server-akka.jar --http-port=80 --https-port=443
+java -jar /trans/target/scala-2.11/trans-server-akka.jar --http-port=80 --https-port=443 "$@"