+++ src/main/scala/io/github/nwtgck/trans_server/Core.scala
@@ -43,7 +43,7 @@ private [this] case class Params(duration       : FiniteDuration,
                                  usesSecureChar : Boolean,
                                  getKeyOpt      : Option[String])
 
-class Core(db: Database, fileDbPath: String)(implicit materializer: ActorMaterializer){
+class Core(db: Database, fileDbPath: String, enableTopPageHttpsRedirect: Boolean)(implicit materializer: ActorMaterializer){
 
   // Logger
   private[this] val logger = Logging.getLogger(materializer.system, this)
@@ -245,10 +245,21 @@ class Core(db: Database, fileDbPath: String)(implicit materializer: ActorMateria
     get {
       // "Get /" for confirming whether the server is running
       pathSingleSlash {
-        // Redirect http to https in Heroku or IBM Cloud (Bluemix)
-        Util.xForwardedProtoHttpsRedirectRoute(
-          getFromResource("index.html") // (from: https://doc.akka.io/docs/akka-http/current/scala/http/routing-dsl/directives/file-and-resource-directives/getFromResource.html)
-        )
+        extractUri { uri =>
+          // If top-page redirect is enable and scheme is HTTP
+          if (enableTopPageHttpsRedirect && uri.scheme == "http") {
+            // Redirect to HTTPs page
+            redirect(
+              uri.copy(scheme = "https"),
+              StatusCodes.PermanentRedirect
+            )
+          } else {
+            // Redirect http to https in Heroku or IBM Cloud (Bluemix)
+            Util.xForwardedProtoHttpsRedirectRoute(
+              getFromResource("index.html") // (from: https://doc.akka.io/docs/akka-http/current/scala/http/routing-dsl/directives/file-and-resource-directives/getFromResource.html)
+            )
+          }
+        }
       } ~
       // Version routing
       path(Setting.GetRouteName.Version) {